+++ src/main/resources/mappers/LoginIdCheckMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.nmfw.foodietree.domain.store.repository.LoginIdCheckMapper">
+<mapper namespace="org.nmfw.foodietree.domain.store.mapper.LoginIdCheckMapper">
     <select id="existsById" resultType="boolean">
         SELECT COUNT(*)
         FROM tbl_store
