+++ src/main/resources/mappers/StoreMapper.xml
@@ -3,7 +3,7 @@
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="org.nmfw.foodietree.domain.store.repository.StoreMapper">
+<mapper namespace="org.nmfw.foodietree.domain.store.mapper.StoreMapper">
 
   <insert id="save">
     INSERT INTO tbl_store
