+++ src/test/java/org/nmfw/foodietree/domain/store/mapper/StoreRepositoryTest.java
