+++ src/test/java/org/nmfw/foodietree/domain/store/service/StoreSignUpServiceTest.java
@@ -3,7 +3,7 @@ package org.nmfw.foodietree.domain.store.service;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.nmfw.foodietree.domain.store.entity.Store;
-import org.nmfw.foodietree.domain.store.repository.StoreMapper;
+import org.nmfw.foodietree.domain.store.mapper.StoreMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.crypto.password.PasswordEncoder;
